+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/FilterSettings.java
@@ -15,11 +15,28 @@ public class FilterSettings {
   private boolean withPupPack;
   private boolean withAltSound;
   private boolean withAltColor;
-  private boolean withPovIni;
+  private boolean withPov;
+  private boolean withIni;
   private boolean withNVOffset;
   private boolean withAlias;
   private int gameStatus = -1;
   private NoteType noteType;
+
+  public boolean isWithPov() {
+    return withPov;
+  }
+
+  public void setWithPov(boolean withPov) {
+    this.withPov = withPov;
+  }
+
+  public boolean isWithIni() {
+    return withIni;
+  }
+
+  public void setWithIni(boolean withIni) {
+    this.withIni = withIni;
+  }
 
   public NoteType getNoteType() {
     return noteType;
@@ -133,14 +150,6 @@ public class FilterSettings {
     this.withAltColor = withAltColor;
   }
 
-  public boolean isWithPovIni() {
-    return withPovIni;
-  }
-
-  public void setWithPovIni(boolean withPovIni) {
-    this.withPovIni = withPovIni;
-  }
-
   public boolean isNotPlayed() {
     return notPlayed;
   }
@@ -187,7 +196,8 @@ public class FilterSettings {
         && !this.withAltColor
         && !this.withAltSound
         && !this.withBackglass
-        && !this.withPovIni
+        && !this.withIni
+        && !this.withPov
         && !this.withPupPack
         && !this.withNVOffset
         && !this.withAlias