+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/frontend/FrontendConnector.java
@@ -8,13 +8,7 @@ import java.util.Map;
 import de.mephisto.vpin.connectors.assets.TableAssetsAdapter;
 import de.mephisto.vpin.restclient.JsonSettings;
 import de.mephisto.vpin.restclient.alx.TableAlxEntry;
-import de.mephisto.vpin.restclient.frontend.Emulator;
-import de.mephisto.vpin.restclient.frontend.Frontend;
-import de.mephisto.vpin.restclient.frontend.FrontendControl;
-import de.mephisto.vpin.restclient.frontend.FrontendControls;
-import de.mephisto.vpin.restclient.frontend.Playlist;
-import de.mephisto.vpin.restclient.frontend.TableDetails;
-import de.mephisto.vpin.restclient.frontend.VPinScreen;
+import de.mephisto.vpin.restclient.frontend.*;
 import de.mephisto.vpin.server.games.Game;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -34,6 +28,11 @@ public interface FrontendConnector {
   File getInstallationFolder();
 
   List<Emulator> getEmulators();
+
+  /**
+   * Force refresh of the whole connector since they can have their own cache, e.g. emulators
+   */
+  void clearCache();
   
   Game getGame(int id);
 
@@ -93,6 +92,8 @@ public interface FrontendConnector {
   boolean isPupPackDisabled(@NonNull Game game);
 
   void setPupPackEnabled(@NonNull Game game, boolean enable);
+
+  List<FrontendPlayerDisplay> getFrontendPlayerDisplays();
 
   //----------------------------------
   // Playlists management