+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/preferences/CardGenerationPreferencesController.java
@@ -4,9 +4,10 @@ import de.mephisto.vpin.commons.fx.Debouncer;
 import de.mephisto.vpin.commons.utils.WidgetFactory;
 import de.mephisto.vpin.restclient.PreferenceNames;
 import de.mephisto.vpin.restclient.cards.CardSettings;
-import de.mephisto.vpin.restclient.highscores.HighscoreCardResolution;
 import de.mephisto.vpin.restclient.frontend.FrontendControl;
+import de.mephisto.vpin.restclient.frontend.FrontendType;
 import de.mephisto.vpin.restclient.frontend.VPinScreen;
+import de.mephisto.vpin.restclient.highscores.HighscoreCardResolution;
 import de.mephisto.vpin.restclient.puppacks.PupPackRepresentation;
 import de.mephisto.vpin.ui.Studio;
 import de.mephisto.vpin.ui.events.EventManager;
@@ -15,6 +16,7 @@ import javafx.application.Platform;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Label;
@@ -27,6 +29,7 @@ import org.slf4j.LoggerFactory;
 import java.awt.*;
 import java.net.URI;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Optional;
 import java.util.ResourceBundle;
@@ -38,7 +41,7 @@ public class CardGenerationPreferencesController implements Initializable {
   public static Debouncer debouncer = new Debouncer();
 
   @FXML
-  private ComboBox<String> popperScreenCombo;
+  private ComboBox<String> cardTargetScreenCombo;
 
   @FXML
   private ComboBox<String> rotationCombo;
@@ -79,15 +82,25 @@ public class CardGenerationPreferencesController implements Initializable {
 
   @Override
   public void initialize(URL url, ResourceBundle resourceBundle) {
-    menuPupPack = client.getPupPackService().getMenuPupPack();
     validationError.managedProperty().bindBidirectional(validationError.visibleProperty());
     transparencyHelp.setVisible(false);
 
+    FrontendType frontendType = client.getFrontendService().getFrontendType();
+
+    ObservableList<String> screenNames = FXCollections.observableList(new ArrayList<>());
+    if (frontendType.equals(FrontendType.Popper)) {
+      menuPupPack = client.getPupPackService().getMenuPupPack();
+      screenNames.addAll("", VPinScreen.Other2.name(), VPinScreen.GameInfo.name(), VPinScreen.GameHelp.name());
+    }
+    if (frontendType.equals(FrontendType.PinballX)) {
+      screenNames.addAll("", VPinScreen.Topper.name(), VPinScreen.DMD.name());
+    }
+
     cardSettings = client.getPreferenceService().getJsonPreference(PreferenceNames.HIGHSCORE_CARD_SETTINGS, CardSettings.class);
 
-    popperScreenCombo.setItems(FXCollections.observableList(Arrays.asList("", VPinScreen.Other2.name(), VPinScreen.GameInfo.name(), VPinScreen.GameHelp.name())));
-    popperScreenCombo.setValue(cardSettings.getPopperScreen() != null ? cardSettings.getPopperScreen() : "");
-    popperScreenCombo.valueProperty().addListener((observable, oldValue, newValue) -> {
+    cardTargetScreenCombo.setItems(screenNames);
+    cardTargetScreenCombo.setValue(cardSettings.getPopperScreen() != null ? cardSettings.getPopperScreen() : "");
+    cardTargetScreenCombo.valueProperty().addListener((observable, oldValue, newValue) -> {
       cardSettings.setPopperScreen(newValue);
       client.getPreferenceService().setJsonPreference(PreferenceNames.HIGHSCORE_CARD_SETTINGS, cardSettings);
       onScreenChange();
@@ -138,40 +151,45 @@ public class CardGenerationPreferencesController implements Initializable {
   }
 
   private void onScreenChange() {
-    String selectedItem = popperScreenCombo.getSelectionModel().getSelectedItem();
+    FrontendType frontendType = client.getFrontendService().getFrontendType();
+    String selectedItem = cardTargetScreenCombo.getSelectionModel().getSelectedItem();
     highscoreCardDuration.setDisable(selectedItem == null);
 
-    if (!StringUtils.isEmpty(selectedItem)) {
-      FrontendControl fn = client.getFrontendService().getPinUPControlFor(VPinScreen.valueOf(selectedItem));
+    validationError.setVisible(false);
 
-      String msg = null;
-      if (fn != null) {
-        if (!fn.isActive()) {
-          msg = "The screen has not been activated in PinUP Popper.";
+    if(frontendType.equals(FrontendType.Popper)) {
+      if (!StringUtils.isEmpty(selectedItem)) {
+        FrontendControl fn = client.getFrontendService().getPinUPControlFor(VPinScreen.valueOf(selectedItem));
+
+        String msg = null;
+        if (fn != null) {
+          if (!fn.isActive()) {
+            msg = "The screen has not been activated in PinUP Popper.";
+          }
+
+          if (fn.getCtrlKey() == 0) {
+            msg = "The screen is not bound to any key in PinUP Popper.";
+          }
         }
 
-        if (fn.getCtrlKey() == 0) {
-          msg = "The screen is not bound to any key in PinUP Popper.";
-        }
+        validationError.setVisible(msg != null);
+        validationError.setText(msg);
+      }
+      else {
+        validationError.setVisible(false);
       }
 
-      validationError.setVisible(msg != null);
-      validationError.setText(msg);
-    }
-    else {
-      validationError.setVisible(false);
-    }
-
-    transparencyHelp.setVisible(false);
-    if (menuPupPack != null && popperScreenCombo.getValue() != null) {
-      String value = popperScreenCombo.getValue();
-      if (!StringUtils.isEmpty(value)) {
-        VPinScreen screen = VPinScreen.valueOf(value);
-        transparencyHelp.setVisible(
-            (screen.equals(VPinScreen.GameHelp) && !menuPupPack.isHelpTransparency()) ||
-                (screen.equals(VPinScreen.GameInfo) && !menuPupPack.isInfoTransparency()) ||
-                (screen.equals(VPinScreen.Other2) && !menuPupPack.isOther2Transparency())
-        );
+      transparencyHelp.setVisible(false);
+      if (menuPupPack != null && cardTargetScreenCombo.getValue() != null) {
+        String value = cardTargetScreenCombo.getValue();
+        if (!StringUtils.isEmpty(value)) {
+          VPinScreen screen = VPinScreen.valueOf(value);
+          transparencyHelp.setVisible(
+              (screen.equals(VPinScreen.GameHelp) && !menuPupPack.isHelpTransparency()) ||
+                  (screen.equals(VPinScreen.GameInfo) && !menuPupPack.isInfoTransparency()) ||
+                  (screen.equals(VPinScreen.Other2) && !menuPupPack.isOther2Transparency())
+          );
+        }
       }
     }