+++ RELEASE_NOTES.md
@@ -1,12 +1,15 @@
 ## Release Notes 2.21.1
 
+- **PinballX Support**:
+- **iScored Subscriptions**:
 - **Notifications**:
-- **Controller Support**: All VPin UI related keybindings have been re-implemeneted using a new API which also includes controller support.   
+- **Controller Support**: All VPin UI related keybindings have been re-implemeneted using a new API which also includes controller support.
 - **Preferences Changes**: 
+- **Filter Section**: Splitted filter preferences for .ini and .pov files.
+
 
 ### Bugfixes
 
 - **Pause Menu**: Fixed auto-play issues of YouTube videos.
-
 - **VPin Studio Launcher**: Improved error handling so that the Studio still launches on seven-zip issues.
 - **VPX Monitor**: I disabled this feature for now since when enabled, this VPX process watcher hangs and I haven't found the problem yet.